move updatePipeline to FC wrapper
